refactor: extract MarketplacePage logic into custom hook
